requirements.txt
Babel==2.7.0
backcall==0.1.0
coverage==4.5.3
crispy-forms-foundation==0.7.1
decorator==4.4.0
Django==2.2.1
django-bootstrap-datepicker==1.3.1
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.7.2
django-filter==2.1.0
django-phonenumber-field==3.0.1
django-tables2==2.0.6
django-tastypie==0.14.2
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
parso==0.4.0
pexpect==4.7.0
phonenumbers==8.10.13
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
python-dateutil==2.8.0
python-mimeparse==1.6.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
wcwidth==0.1.7
